requirements.txt
fastapi==0.118.0
uvicorn[standard]==0.37.0
pydantic-settings==2.11.0
PyJWT[crypto]==2.10.1
SQLAlchemy==2.0.43
httpx==0.28.1
psycopg[binary]==3.2.9
